refactor: move `getMarketingCta` to common/users
