+++ src/routes/boot.ts
@@ -1,11 +1,6 @@
-import {
-  FastifyBaseLogger,
-  FastifyInstance,
-  FastifyReply,
-  FastifyRequest,
-} from 'fastify';
+import { FastifyInstance, FastifyReply, FastifyRequest } from 'fastify';
 import createOrGetConnection from '../db';
-import { DataSource, EntityManager, IsNull } from 'typeorm';
+import { DataSource, EntityManager } from 'typeorm';
 import { clearAuthentication, dispatchWhoami } from '../kratos';
 import { generateUUID } from '../ids';
 import { generateSessionId, setTrackingId } from '../tracking';
@@ -21,7 +16,6 @@ import {
   SourceMember,
   SquadSource,
   User,
-  UserMarketingCta,
 } from '../entity';
 import {
   getPermissionsForMember,
@@ -34,18 +28,16 @@ import {
   getRedisObject,
   ioRedisPool,
   ONE_DAY_IN_SECONDS,
-  ONE_WEEK_IN_SECONDS,
-  RedisMagicValues,
   setRedisObject,
   setRedisObjectWithExpiry,
 } from '../redis';
+import { generateStorageKey, REDIS_BANNER_KEY, StorageTopic } from '../config';
 import {
-  generateStorageKey,
-  REDIS_BANNER_KEY,
-  StorageKey,
-  StorageTopic,
-} from '../config';
-import { base64, getSourceLink, submitArticleThreshold } from '../common';
+  base64,
+  getMarketingCta,
+  getSourceLink,
+  submitArticleThreshold,
+} from '../common';
 import { AccessToken, signJwt } from '../auth';
 import { cookies, setCookie, setRawCookie } from '../cookies';
 import { parse } from 'graphql/language/parser';
@@ -341,49 +333,6 @@ const getUser = (con: DataSource, userId: string): Promise<User> =>
       'cover',
     ],
   });
-
-export const getMarketingCta = async (
-  con: DataSource,
-  log: FastifyBaseLogger,
-  userId: string,
-) => {
-  if (!userId) {
-    log.info('no userId provided when fetching marketing cta');
-    return null;
-  }
-
-  const redisKey = generateStorageKey(
-    StorageTopic.Boot,
-    StorageKey.MarketingCta,
-    userId,
-  );
-  const rawRedisValue = await getRedisObject(redisKey);
-
-  if (rawRedisValue === RedisMagicValues.SLEEPING) {
-    return null;
-  }
-
-  // If the vale in redis is `EMPTY`, we fallback to `null`
-  let marketingCta: MarketingCta | null = JSON.parse(rawRedisValue || null);
-
-  // If the key is not in redis, we need to fetch it from the database
-  if (!marketingCta) {
-    const userMarketingCta = await con.getRepository(UserMarketingCta).findOne({
-      where: { userId, readAt: IsNull() },
-      order: { createdAt: 'ASC' },
-      relations: ['marketingCta'],
-    });
-
-    marketingCta = userMarketingCta?.marketingCta || null;
-    const redisValue = userMarketingCta
-      ? JSON.stringify(marketingCta)
-      : RedisMagicValues.SLEEPING;
-
-    await setRedisObjectWithExpiry(redisKey, redisValue, ONE_WEEK_IN_SECONDS);
-  }
-
-  return marketingCta;
-};
 
 const loggedInBoot = async (
   con: DataSource,
