+++ src/schema/users.ts
@@ -41,6 +41,7 @@ import {
   GQLUserStreakTz,
   toGQLEnum,
   getUserPermalink,
+  getMarketingCta,
 } from '../common';
 import { getSearchQuery, GQLEmptyResponse, processSearchQuery } from './common';
 import { ActiveView } from '../entity/ActiveView';
@@ -60,7 +61,6 @@ import { getTimezoneOffset } from 'date-fns-tz';
 import { markdown } from '../common/markdown';
 import { deleteRedisKey } from '../redis';
 import { StorageKey, StorageTopic, generateStorageKey } from '../config';
-import { getMarketingCta } from '../routes/boot';
 
 export interface GQLUpdateUserInput {
   name: string;
